Tool can now map metadata from Drive to Box
